--- config/config.json
@@ -1,23 +1,0 @@
-{
-  "development": {
-    "username": "root",
-    "password": null,
-    "database": "ddas5d1bba36av",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  },
-  "test": {
-    "username": "root",
-    "password": null,
-    "database": "ddas5d1bba36av",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  },
-  "production": {
-    "username": "root",
-    "password": null,
-    "database": "ddas5d1bba36av",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  }
-}
