+++ controllers/catsController.js
@@ -1,0 +1,50 @@
+var express = require("express");
+
+var router = express.Router();
+
+// Import the model (cat.js) to use its database functions.
+var cat = require("../models/cat.js");
+
+// Create all our routes and set up logic within those routes where required.
+router.get("/", function(req, res) {
+  cat.all(function(data) {
+    var hbsObject = {
+      cats: data
+    };
+    console.log(hbsObject);
+    res.render("index", hbsObject);
+  });
+});
+
+router.post("/", function(req, res) {
+  cat.create([
+    "name", "sleepy"
+  ], [
+    req.body.name, req.body.sleepy
+  ], function() {
+    res.redirect("/");
+  });
+});
+
+router.put("/:id", function(req, res) {
+  var condition = "id = " + req.params.id;
+
+  console.log("condition", condition);
+
+  cat.update({
+    sleepy: req.body.sleepy
+  }, condition, function() {
+    res.redirect("/");
+  });
+});
+
+router.delete("/:id", function(req, res) {
+  var condition = "id = " + req.params.id;
+
+  cat.delete(condition, function() {
+    res.redirect("/");
+  });
+});
+
+// Export routes for server.js to use.
+module.exports = router;
